requirements.txt
#pip install -r requirements.txt
Flask
gunicorn
google-cloud-firestore
google-cloud-storage
google-cloud-pubsub
google-cloud-logging
Werkzeug
python-dotenv
wave